+++ react/components/molecules/chat-message/chat-message.js
@@ -1,11 +1,12 @@
 import React, { PureComponent } from "react";
-import { StyleSheet, Text as RNText, View, ViewPropTypes } from "react-native";
+import { StyleSheet, View, ViewPropTypes } from "react-native";
 import PropTypes from "prop-types";
 import { dateTimeString, isImage } from "ripe-commons-native";
 
 import { baseStyles } from "../../../util";
 
-import { Attachment, Avatar, Lightbox, Text } from "../../atoms";
+import { Attachment, Avatar, Lightbox, StatusEntry, Text } from "../../atoms";
+import { AvatarList } from "../avatar-list";
 
 export class ChatMessage extends PureComponent {
     static get propTypes() {
@@ -13,8 +14,12 @@ export class ChatMessage extends PureComponent {
             id: PropTypes.number,
             avatarUrl: PropTypes.string.isRequired,
             username: PropTypes.string.isRequired,
-            message: PropTypes.string.isRequired,
             date: PropTypes.number.isRequired,
+            message: PropTypes.string,
+            status: PropTypes.string,
+            statusProps: PropTypes.object,
+            replies: PropTypes.number,
+            repliesAvatars: PropTypes.array,
             attachments: PropTypes.arrayOf(
                 PropTypes.exact({
                     name: PropTypes.string.isRequired,
@@ -31,8 +36,12 @@ export class ChatMessage extends PureComponent {
         return {
             avatarUrl: undefined,
             username: undefined,
-            message: undefined,
             date: undefined,
+            message: undefined,
+            status: undefined,
+            statusProps: {},
+            replies: undefined,
+            repliesAvatars: [],
             attachments: [],
             imagePlaceholder: undefined,
             style: {},
@@ -52,15 +61,30 @@ export class ChatMessage extends PureComponent {
         return (
             <View style={[styles.chatMessage, this.props.style]}>
                 <Avatar style={styles.avatar} image={{ uri: this.props.avatarUrl }} size={32} />
-                <View>
+                <View style={styles.message}>
                     <View style={styles.header}>
-                        <RNText style={styles.username}>{this.props.username}</RNText>
-                        <RNText style={styles.date}>
+                        <Text style={styles.username}>{this.props.username}</Text>
+                        <Text style={styles.date}>
                             {dateTimeString(this.props.date, { seconds: false })}
-                        </RNText>
+                        </Text>
                     </View>
                     {this.props.message ? (
                         <Text style={styles.text}>{this.props.message}</Text>
+                    ) : null}
+                    {this.props.status ? (
+                        <StatusEntry
+                            style={styles.status}
+                            status={this.props.status}
+                            {...this.props.statusProps}
+                        />
+                    ) : null}
+                    {this.props.replies ? (
+                        <View style={styles.replies}>
+                            <AvatarList avatars={this.props.repliesAvatars} size={24} />
+                            <Text style={styles.repliesText}>
+                                {this.props.replies} {this.props.replies > 1 ? "replies" : "reply"}
+                            </Text>
+                        </View>
                     ) : null}
                     {this.props.attachments.map((attachment, index) => (
                         <View style={this._attachmentsStyle()} key={index}>
@@ -76,7 +100,7 @@ export class ChatMessage extends PureComponent {
                                 <Attachment
                                     style={this._attachmentStyle(index)}
                                     filename={attachment.name}
-                                    url={attachment.url}
+                                    url={attachment.path}
                                 />
                             )}
                         </View>
@@ -90,20 +114,19 @@ export class ChatMessage extends PureComponent {
 const styles = StyleSheet.create({
     chatMessage: {
         flexDirection: "row",
-        marginStart: 15,
-        marginEnd: 15
+        width: "100%"
     },
     avatar: {
-        marginEnd: 13,
+        marginEnd: 9,
         marginTop: 3
     },
-    header: {
-        flexDirection: "row",
-        marginTop: 3,
-        alignItems: "center"
+    message: {
+        flex: 1
     },
-    text: {
-        marginEnd: 40
+    header: {
+        justifyContent: "flex-start",
+        flexDirection: "column",
+        marginTop: 3
     },
     username: {
         fontFamily: baseStyles.FONT_BOLD,
@@ -117,6 +140,23 @@ const styles = StyleSheet.create({
         color: "#a4adb5",
         fontSize: 11,
         lineHeight: 18
+    },
+    text: {
+        marginTop: 4
+    },
+    status: {
+        marginTop: 4
+    },
+    replies: {
+        marginTop: 8,
+        flexDirection: "row",
+        alignItems: "center"
+    },
+    repliesText: {
+        marginLeft: 8,
+        fontFamily: baseStyles.FONT,
+        fontSize: 14,
+        color: "#6051f2"
     }
 });