+++ react/components/organisms/chat/chat.stories.js
@@ -105,10 +105,47 @@ storiesOf("Organisms", module)
                 attachments: []
             },
             {
+                avatarUrl: "https://id.platforme.com/admin/accounts/ns%40platforme.com/avatar",
+                username: "NFSS10",
+                status: "created",
+                replies: 5,
+                repliesAvatars: [
+                    "https://i.pravatar.cc",
+                    "https://i.pravatar.cc?1",
+                    "https://i.pravatar.cc?2",
+                    "https://i.pravatar.cc?3",
+                    "https://i.pravatar.cc?4"
+                ],
+                date: 1574950742823,
+                attachments: []
+            },
+            {
                 avatarUrl: "https://id.platforme.com/admin/accounts/t-ms%40platforme.com/avatar",
                 username: "BeeMargarida",
                 message: "Donec varius eget eros a euismod.",
                 date: 1574950742823,
+                attachments: []
+            },
+            {
+                avatarUrl: "https://id.platforme.com/admin/accounts/t-ms%40platforme.com/avatar",
+                username: "BeeMargarida",
+                message: "Lorem ipsum dolor sit.",
+                date: 1574950742823,
+                attachments: []
+            },
+            {
+                avatarUrl: "https://id.platforme.com/admin/accounts/t-ms%40platforme.com/avatar",
+                username: "BeeMargarida",
+                message:
+                    "Sed tristique faucibus interdum. Cras egestas efficitur sem id mollis. Cras varius malesuada orci, in vulputate urna vulputate a.",
+                date: 1574950746423,
+                attachments: []
+            },
+            {
+                avatarUrl: "https://id.platforme.com/admin/accounts/t-ms%40platforme.com/avatar",
+                username: "BeeMargarida",
+                message: "Lorem ipsum.",
+                date: 1574950746424,
                 attachments: []
             }
         ];